feat: APIs for user demo; demo_user.py; demo_user_feedback.py; EvaluateAnswerRequest changed eval_distribution List[int] to List[float]

+++ src/schemas/endpoints/schema.py
@@ -22,7 +22,7 @@ class EvaluateAnswerRequest(BaseModel):
     question: str
     interview_id: int
     answer: str
-    eval_distribution: List[int] = Field(default_factory=list) #[0,0,0,0,0,0,0]
+    eval_distribution: List[float] = Field(default_factory=list) #[0,0,0,0,0,0,0]
 
 class GenerateHintRequest(BaseModel):
     question : str
